feat: separate document loading and retrieval into distinct classes
