+++ apps/api/models/data_loader/__init__.py
@@ -1,0 +1,1 @@
+from .PDF import PDFLoader
