+++ apps/api/models/retrieval/models/PDF.py
@@ -12,15 +12,8 @@ from langchain.embeddings.openai import OpenAIEmbeddings
 from langchain.llms import OpenAI
 from langchain.retrievers.self_query.base import SelfQueryRetriever
 from langchain.schema import Document
-from langchain.text_splitter import CharacterTextSplitter
 from langchain.vectorstores import Pinecone
-from pdfminer.converter import TextConverter
-from pdfminer.layout import LAParams
-from pdfminer.pdfdocument import PDFDocument
-from pdfminer.pdfinterp import PDFPageInterpreter, PDFResourceManager
-from pdfminer.pdfpage import PDFPage
-from pdfminer.pdfparser import PDFParser
-from pydantic import BaseModel, Field
+from models.data_loader import PDFLoader
 from utils import PINECONE_API_KEY, PINECONE_ENVIRONMENT
 from pdfminer.converter import TextConverter
 from pdfminer.layout import LAParams
@@ -65,50 +58,6 @@ def extract_text_from_pdf(contents: io.BytesIO) -> list:
 
 
 logger = logging.getLogger(__name__)
-
-
-def extract_text_from_pdf(contents: io.BytesIO) -> list:
-    resource_manager = PDFResourceManager()
-    fake_file_handle = io.StringIO()
-    converter = TextConverter(resource_manager, fake_file_handle, laparams=LAParams())
-    page_interpreter = PDFPageInterpreter(resource_manager, converter)
-    for page in PDFPage.get_pages(contents, caching=True, check_extractable=True):
-        page_interpreter.process_page(page)
-    text = fake_file_handle.getvalue()
-    pages = text.split("\f")
-
-    # Remove the last line of each page if it's a number or its length is less than 5
-    for i in range(len(pages)):
-        lines = pages[i].split("\n")
-        if len(lines) > 1:  # Ensure there is more than one line
-            last_line = lines[-1]
-            if last_line.isdigit() or len(last_line) < 5:
-                logger.debug(f"Removing last line: {last_line}")
-                lines = lines[:-1]  # Remove the last line
-                pages[i] = "\n".join(lines)
-
-    # Join the pages back together
-    text = "\f".join(pages)
-
-    converter.close()
-    fake_file_handle.close()
-
-    return text
-
-
-class PDFSplitterOption(BaseModel):
-    type: str = Field(default="character")
-    chunk_size: int = Field(default=1000)
-    chunk_overlap: int = Field(default=0)
-
-
-class PDFEmbeddingOption(BaseModel):
-    model: str = Field(default="gpt-3.5-turbo")
-
-
-class PDFRetrivalOption(BaseModel):
-    splitter: PDFSplitterOption = Field(default_factory=PDFSplitterOption)
-    embedding: PDFEmbeddingOption = Field(default_factory=PDFEmbeddingOption)
 
 
 class PatchedSelfQueryRetriever(SelfQueryRetriever):
@@ -132,50 +81,13 @@ class PatchedSelfQueryRetriever(SelfQueryRetriever):
             )
         else:
             raise ValueError(f"search_type of {self.search_type} not allowed.")
+
         return docs
 
 
 class PDFRetrieverMixin:
-    def load_and_split_documents(self):
-        doc = []
-        for dataset in self.datasets:
-            logger.info(f"Loading dataset {dataset.id}")
-            _doc = []
-            options = PDFRetrivalOption(**dataset.retrieval)
-
-            text_splitter = CharacterTextSplitter.from_tiktoken_encoder(
-                separator=" ",
-                chunk_size=options.splitter.chunk_size,
-                chunk_overlap=options.splitter.chunk_overlap,
-            )
-
-            for document in dataset.documents:
-                if document.type == "pdf":
-                    pdf_content = self.storage_client.load(document.url)
-                    text = extract_text_from_pdf(pdf_content)
-                    pages = text.split("\f")
-                    for page_number, page in enumerate(pages):
-                        _doc.append(
-                            Document(
-                                page_content=page,
-                                metadata={
-                                    "source": document.url,
-                                    "page_number": page_number,
-                                },
-                            )
-                        )
-                else:
-                    logger.error(f"Document type {document.type} not supported")
-                    raise Exception("Document type not supported")
-            _doc = text_splitter.split_documents(_doc)
-            logger.info(
-                f"got documents: {len(_doc)} while loading dataset {dataset.id}"
-            )
-            doc += _doc
-        return doc
-
     def create_retriever(self):
-        doc = self.load_and_split_documents()
+        doc = PDFLoader.load_and_split_documents(self.dataset_ids)
 
         embeddings = OpenAIEmbeddings(model="text-embedding-ada-002")
         pinecone.init(api_key=PINECONE_API_KEY, environment=PINECONE_ENVIRONMENT)
